requirements.txt
anyio
appnope
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
attrs
Babel
beautifulsoup4
bleach
certifi
cffi
charset-normalizer
comm
contourpy
cycler
debugpy
decorator
defusedxml
executing
fastjsonschema
filelock
fonttools
fqdn
fsspec
idna
imageio
ipykernel
ipython
ipywidgets
isoduration
jedi
Jinja2
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter_client
jupyter-console
jupyter_core
jupyter-events
jupyter-lsp
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab-widgets
kiwisolver
lazy_loader
MarkupSafe
matplotlib
matplotlib-inline
mistune
mpmath
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numpy
opencv-python
overrides
packaging
pandas
pandocfilters
parso
pexpect
Pillow
pip
platformdirs
prometheus-client
prompt-toolkit
psutil
ptyprocess
pure-eval
pycparser
Pygments
pyparsing
python-dateutil
python-json-logger
pytz
PyWavelets
PyYAML
pyzmq
qtconsole
QtPy
referencing
requests
rfc3339-validator
rfc3986-validator
rpds-py
scikit-image
scikit-learn
scipy
seaborn
Send2Trash
setuptools
six
sniffio
soupsieve
stack-data
sympy
terminado
threadpoolctl
tifffile
tinycss2
torch
torchaudio
torchvision
tornado
tqdm
traitlets
types-python-dateutil
typing_extensions
tzdata
uri-template
urllib3
wcwidth
webcolors
webencodings
websocket-client
wheel
widgetsnbextension
